Fix rendering issue on playgrounds
