+++ src/components/image/component.js
@@ -1,23 +1,44 @@
 import React from "react"
 import PropTypes from "prop-types"
 import Img from "gatsby-image"
-import classNames from "classnames/bind"
+import classNames from "classnames"
 import dig from "object-dig"
+
 import styles from "./styles.module.scss"
 
 const Image = ({ alt, className, src, ...props }) => {
-  const classes = classNames(styles.cl_image, { [className]: className })
+  const defaultAltAttribute = image =>
+    image
+      .split("/")
+      .pop()
+      .split(".")
+      .slice(0, -1)
+      .join(".")
+
+  const classes = classNames(styles.image, { [className]: className })
 
   // ---------------------------------------- | Gastby Image
 
   if (dig(src, "childImageSharp", "fluid") || dig(src, "childImageSharp", "fixed")) {
-    return <Img className={classes} alt={alt} {...src.childImageSharp} />
+    let imageName = dig(src, "childImageSharp", "fluid", "src")
+
+    if (dig(src, "childImageSharp", "fixed")) {
+      imageName = dig(src, "childImageSharp", "fixed", "src")
+    }
+
+    return (
+      <Img
+        className={classes}
+        alt={alt || defaultAltAttribute(imageName)}
+        {...src.childImageSharp}
+      />
+    )
   }
 
   // ---------------------------------------- | Native Image
 
   if (typeof src === "string") {
-    return <img className={classes} src={src} alt={alt} {...props} />
+    return <img className={classes} src={src} alt={alt || defaultAltAttribute(src)} {...props} />
   }
 
   // ---------------------------------------- | Invalid src
